requirements.txt
Telethon==1.38.1
tqdm==4.67.1
colorama==0.4.6
cryptg==0.5.0.post0
python-dotenv==1.0.1
customtkinter==5.2.2
pillow==10.3.0
PySocks==1.7.1
PySide6>=6.5.0
humanize>=4.0.0